requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
passlib[bcrypt]
python-jose[cryptography]
pydantic
tensorflow==2.15.0
numpy
opencv-python
